+++ alioth/src/loader/linux/x86_64.rs
@@ -42,7 +42,7 @@ const MINIMAL_VERSION: u16 = 0x020c;
 
 pub fn load<P: AsRef<Path>>(
     memory: &RamBus,
-    mem_regions: &[(usize, MemRegionEntry)],
+    mem_regions: &[(u64, MemRegionEntry)],
     kernel: P,
     cmd_line: Option<&str>,
     initramfs: Option<P>,
@@ -92,11 +92,15 @@ pub fn load<P: AsRef<Path>>(
     // load cmd line
     if let Some(cmd_line) = cmd_line {
         let cmd_line_limit =
-            std::cmp::min(boot_params.hdr.cmdline_size as usize, KERNEL_CMD_LINE_LIMIT);
-        if cmd_line.len() > cmd_line_limit {
+            std::cmp::min(boot_params.hdr.cmdline_size as u64, KERNEL_CMD_LINE_LIMIT);
+        if cmd_line.len() as u64 > cmd_line_limit {
             return Err(Error::CmdLineTooLong(cmd_line.len(), cmd_line_limit));
         }
-        memory.write_range(KERNEL_CMD_LINE_START, cmd_line.len(), cmd_line.as_bytes())?;
+        memory.write_range(
+            KERNEL_CMD_LINE_START,
+            cmd_line.len() as u64,
+            cmd_line.as_bytes(),
+        )?;
         boot_params.hdr.cmd_line_ptr = KERNEL_CMD_LINE_START as u32;
         boot_params.ext_cmd_line_ptr = (KERNEL_CMD_LINE_START >> 32) as u32;
     }
@@ -104,18 +108,18 @@ pub fn load<P: AsRef<Path>>(
     // load kernel image
     let kernel_offset = (boot_params.hdr.setup_sects as u64 + 1) * 512;
     kernel.seek(SeekFrom::Start(kernel_offset))?;
-    let kernel_size = (kernel_meta.len() - kernel_offset) as usize;
+    let kernel_size = kernel_meta.len() - kernel_offset;
     memory.write_range(KERNEL_IMAGE_START, kernel_size, kernel)?;
 
     // load initramfs
     let initramfs_range;
     if let Some(initramfs) = initramfs {
         let initramfs = File::open(initramfs)?;
-        let initramfs_size = initramfs.metadata()?.len() as usize;
+        let initramfs_size = initramfs.metadata()?.len();
         let initramfs_gpa = search_initramfs_address(
             mem_regions,
             initramfs_size,
-            boot_params.hdr.initrd_addr_max as usize,
+            boot_params.hdr.initrd_addr_max as u64,
         )?;
         let initramfs_end = initramfs_gpa + initramfs_size;
         memory.write_range(initramfs_gpa, initramfs_size, initramfs)?;
@@ -144,36 +148,36 @@ pub fn load<P: AsRef<Path>>(
             MemRegionType::Hidden => continue,
         };
         boot_params.e820_table[region_index] = BootE820Entry {
-            addr: *addr as u64,
-            size: region.size as u64,
+            addr: *addr,
+            size: region.size,
             type_,
         };
         region_index += 1;
     }
     boot_params.e820_entries = mem_regions.len() as u8;
 
-    boot_params.acpi_rsdp_addr = EBDA_START as u64;
+    boot_params.acpi_rsdp_addr = EBDA_START;
 
     memory.write(LINUX_BOOT_PARAMS_START, &boot_params)?;
 
     // set up identity paging
     let pml4_start = BOOT_PAGING_START;
     let pdpt_start = pml4_start + 0x1000;
-    let pml4e = (Entry::P | Entry::RW).bits() as u64 | pdpt_start as u64;
+    let pml4e = (Entry::P | Entry::RW).bits() as u64 | pdpt_start;
     memory.write(pml4_start, &pml4e)?;
-    let alignment = boot_params.hdr.kernel_alignment as usize;
+    let alignment = boot_params.hdr.kernel_alignment as u64;
     let runtime_start = (KERNEL_IMAGE_START + alignment - 1) & !(alignment - 1);
     let max_addr = std::cmp::max(
-        runtime_start + boot_params.hdr.init_size as usize,
+        runtime_start + boot_params.hdr.init_size as u64,
         std::cmp::max(
-            LINUX_BOOT_PARAMS_START + size_of::<BootParams>(),
+            LINUX_BOOT_PARAMS_START + size_of::<BootParams>() as u64,
             KERNEL_CMD_LINE_START + KERNEL_CMD_LINE_LIMIT,
         ),
     );
     let num_page = (max_addr as u64 + (1 << 30) - 1) >> 30;
     for i in 0..num_page {
         let pdpte = (i << 30) | (Entry::P | Entry::RW | Entry::PS).bits() as u64;
-        memory.write(pdpt_start + i as usize * size_of::<u64>(), &pdpte)?;
+        memory.write(pdpt_start + i * size_of::<u64>() as u64, &pdpte)?;
     }
 
     // set up gdt
@@ -210,7 +214,7 @@ pub fn load<P: AsRef<Path>>(
         boot_ldtr.to_desc(),
     ];
     let gdtr = DtRegVal {
-        base: BOOT_GDT_START as u64,
+        base: BOOT_GDT_START,
         limit: size_of_val(&gdt) as u16 - 1,
     };
     let idtr = DtRegVal { base: 0, limit: 0 };
@@ -218,14 +222,14 @@ pub fn load<P: AsRef<Path>>(
 
     Ok(InitState {
         regs: vec![
-            (Reg::Rsi, LINUX_BOOT_PARAMS_START as u64),
-            (Reg::Rip, KERNEL_IMAGE_START as u64 + 0x200),
+            (Reg::Rsi, LINUX_BOOT_PARAMS_START),
+            (Reg::Rip, KERNEL_IMAGE_START + 0x200),
             (Reg::Rflags, Rflags::RESERVED_1.bits() as u64),
         ],
         sregs: vec![
             (SReg::Efer, (Efer::LMA | Efer::LME).bits() as u64),
             (SReg::Cr0, (Cr0::NE | Cr0::PE | Cr0::PG).bits() as u64),
-            (SReg::Cr3, pml4_start as u64),
+            (SReg::Cr3, pml4_start),
             (SReg::Cr4, Cr4::PAE.bits() as u64),
         ],
         seg_regs: vec![
